// packages/single-responsibility-principle/src/types/User.ts
export interface User {
  id: number;
  name: string;
  username: string;
  email: string;
  contact: string;
  address: {
    street: string;
    suite: string;
    city: string;
    zipcode: string;
    geo: {
      lat: string;
      lng: string;
    };
  };
  phone: string;
  website: string;
  company: {
    name: string;
    catchPhrase: string;
    bs: string;
  };
}